fixed events functionality
