+++ client/src/Screens/HomeScreen/index.js
@@ -20,14 +20,17 @@ import {
 	GetAsync,
 	CreateAsync,
 } from "../../Components"
+import moment from "moment"
 
 export default function Home({ navigation }) {
 	const [createEvent, setCreateEvent] = useState(false)
 	const [joinEvent, setJoinEvent] = useState(false)
 	const [viewEvent, setViewEvent] = useState(false)
+	const [events, setEvents] = useState([]);
 
 	useEffect(() => {
 		setPreferenceTokens()
+		getEvents().then((events) => setEvents(events));
 	}, [])
 
 	async function clickCreateEvent() {
@@ -38,20 +41,7 @@ export default function Home({ navigation }) {
 		setViewEvent(!joinEvent)
 	}
 
-	const events = [
-		{
-			title: "Event 1",
-			date: "2023-12-01",
-			time: "12:00 PM",
-			location: "Venue 1",
-		},
-		{
-			title: "Event 2",
-			date: "2023-12-01",
-			time: "2:30 PM",
-			location: "Venue 2",
-		},
-	]
+	
 
 	return (
 		// <ImageBackground source={image} style={{ width: "100%", height: "100%" }}>
@@ -69,9 +59,9 @@ export default function Home({ navigation }) {
 					/>
 				) : null}
 			</View>
-			<ViewEventsCarousel  title = {'My events'}events={events} />
+			<ViewEventsCarousel  title = {'My events'} events={events} />
 
-			<ViewEventsCarousel  title = {'My invites'}events={events} />
+			<ViewEventsCarousel  title = {'My invites'} events={events} />
 			{/* <View>
 				<AppButton title="Join Event" onPress={clickJoinEvent} />
 				<Modal
@@ -163,3 +153,35 @@ async function setPreferenceTokens() {
 		console.log("Walk Speed", await GetAsync("walkingSpeed"))
 	}
 }
+async function getEvents() {
+	const token = await GetAsync("token")
+	try {
+		const options = {
+			method: "GET",
+			headers: {
+				"Content-Type": "application/json",
+				"Authorization": token,
+			},
+		}
+		const response = await fetch(
+			"https://metro-mingle.onrender.com/event/all",
+			options
+		)
+
+		const data = await response.json()
+
+		if (response.status == 200) {
+			return filterEventsHappeningToday(data.events)
+		} else {
+			console.log("error")
+		}
+	} catch (error) {
+		console.log(error)
+	}
+}
+
+	//filter for today event
+  const filterEventsHappeningToday = (events) => {
+    const today = moment().format('YYYY-MM-DD');
+    return events.filter(event => moment(event.date).isSame(today, 'day'));
+  };
